refactor market.py before adding async
